rename to plural of constant
